Navlink

+++ migrations/20200916145347_meetings.js
@@ -2,7 +2,7 @@ exports.up = function(knex) {
   return knex.schema.createTable('meetings', table => {
     table.increments('id')
     table.string('meeting_name')
-    table.decimal('duration')
+    table.integer('duration')
     table.integer('attendees')
     table.decimal('cost')
     table.timestamps(true,true)
